requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
nuitka>=4.0.2
numpy>=2.4.2